requirements.txt
aioamqp==0.14.0
aiohttp==3.7.4.post0
amqp==2.6.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.9.3
bidict==0.21.3
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.1
certifi==2020.11.8
cffi==1.14.4
channels==3.0.3
channels-rabbitmq==1.3.1
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cycler==0.10.0
daphne==3.0.1
Django==3.1.4
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-cors-headers==3.6.0
django-currentuser==0.5.2
django-debug-toolbar==3.2.1
django-eventstream==3.1.0
django-extensions==3.1.3
django-filter==2.4.0
django-gm2m==1.1.1
django-grip==3.0.0
django-rabbitmq==0.1.3
django-timezone-field==4.1.1
djangorestframework==3.12.2
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
flake8==3.8.4
flower==0.9.5
gripcontrol==4.0.0
gunicorn==20.0.4
h11==0.12.0
humanize==3.2.0
hyperlink==20.0.1
idna==2.10
idna-ssl==1.1.0
importlib-metadata==1.7.0
incremental==17.5.0
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.1
kombu==4.6.11
lxml==4.6.2
MarkupSafe==2.0.1
matplotlib==3.3.3
mccabe==0.6.1
msgpack==1.0.2
multidict==5.1.0
numpy==1.19.4
pamqp==2.3.0
pandas==1.1.5
pika==1.1.0
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pubcontrol==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-socketio==5.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-docx==0.8.10
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2020.4
pywebhdfs==0.4.1
requests==2.25.0
service-identity==18.1.0
six==1.15.0
socketIO-client==0.7.2
soupsieve==2.1
sqlparse==0.4.1
textblob==0.15.3
tornado==6.1
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.13.3
vine==1.3.0
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==0.16.1
wordcloud==1.8.1
yarn-api-client==1.0.2
yarl==1.6.3
zipp==3.4.0
zope.interface==5.2.0
